Move and rename evaluation-related files
